Main character and Interactable NPC
1. set basic parameter of Main character
2. Main character can detect distance to other NPCs and change status in real time

+++ Social_Distance/Source/Social_Distance/NPC_Interactable.h
@@ -17,16 +17,16 @@ public:
 	ANPC_Interactable();
 
 	UPROPERTY()
-		int32 Loneliness;
+		float Loneliness;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = CharacterProperty)
-		int32 InitLoneliness;
+		float InitLoneliness;
 
 	UPROPERTY()
-		int32 Risk;
+		float Risk;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = CharacterProperty)
-		int32 InitRisk;
+		float InitRisk;
 
 	UPROPERTY(BlueprintReadOnly)
 		FVector SelfLocation;
